:truck: `Collection` 페이지 ui 관련 파일 디렉토리 정리
